+++ CountlyCrashReporter.h
@@ -10,11 +10,12 @@
 #if TARGET_OS_IOS
 @property (nonatomic) BOOL isEnabledOnInitialConfig;
 @property (nonatomic) NSDictionary* crashSegmentation;
+@property (nonatomic) NSUInteger crashLogLimit;
 
 + (instancetype)sharedInstance;
 - (void)startCrashReporting;
 - (void)stopCrashReporting;
-- (void)recordHandledException:(NSException *)exception withStackTrace:(NSArray *)stackTrace;
+- (void)recordException:(NSException *)exception withStackTrace:(NSArray *)stackTrace isFatal:(BOOL)isFatal;
 - (void)log:(NSString *)log;
 #endif
 @end